pyproject.toml
[project]
name = "OpenAI_FireCrawl"
version = "0.0.3"
description = "OpenAI_FireCrawl"
authors = [{ name = "Abhishek Jagtap", email = "Abhijagtap13@gmail.com" }]
dependencies = [
    "uipath-openai-agents>=0.0.1, <0.1.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.11"
